feat: support schema definitions in TypeScript/JavaScript files (#561)
## Summary

- Create `graphql-types` foundation crate consolidating shared types
(`SourceSpan`, `Position`, `Range`, etc.)
- Replace `FileKind` enum with separate `Language` + `DocumentKind`
enums
- Use config patterns (schema vs documents) to determine `DocumentKind`
instead of content inspection
- Add content validation to detect mismatches and fail fast
- Change block URIs from `#block{idx}` to `#L{start}-L{end}` format
- Use `Parser::source_offset()` from apollo-compiler fork for correct
diagnostic positions in TS/JS schema files
- Introduce `SourceSpan` type that bundles byte offsets with block
context, making incorrect lint diagnostic positions structurally
impossible
- Refactor all 9 lint rules and consumers to use `SourceSpan` via
`DocumentRef::span()`, removing the error-prone `with_block_context()`
opt-in pattern

## Details

This PR implements proper support for schema definitions in
TypeScript/JavaScript files and hardens the diagnostic position API.

**Key changes:**

1. **New `graphql-types` crate** - Zero-dependency foundation crate with
`Language`, `DocumentKind`, `FileId`, `Position`, `Range`, `SourceSpan`,
and `Severity` types

2. **Separate Language and DocumentKind** - `Language` (from file
extension) determines parsing strategy; `DocumentKind` (from config)
determines semantic processing

3. **Config-driven classification** - LSP uses config patterns to
determine if a file is schema or executable, not content inspection

4. **Content mismatch validation** - CLI fails fast if schema config
files contain operations or document config files contain type
definitions

5. **Line-range block URIs** - Multi-block TS/JS files now use
`#L15-L42` format instead of `#block0` for better error attribution

6. **Correct diagnostic positions via `SourceSpan`** - Schema validation
diagnostics use `Parser::source_offset()` to bake extraction offsets
into the apollo-compiler parser. For lint diagnostics, the new
`SourceSpan` type bundles byte offsets with block context (line_offset,
byte_offset, source text), and `DocumentRef::span()` automatically fills
in block context. This replaces the old `with_block_context()` opt-in on
`LintDiagnostic` -- forgetting to call it caused the `unused_fields`
position bug. Now it's impossible to create a lint diagnostic without
correct block context.

**SourceSpan refactoring details:**

- `LintDiagnostic` now has a single `span: SourceSpan` field instead of
separate `offset_range`, `block_line_offset`, `block_byte_offset`,
`block_source` fields
- All 6 per-file rules pass `&DocumentRef` through helper functions and
call `doc.span(start, end)`
- All 3 project-wide rules (`unused_fields`, `unused_fragments`,
`unique_names`) store `SourceSpan` in their intermediate structs
- Consumers (`lint_integration.rs`, `server.rs`, `fix.rs`) read from
`ld.span` fields

## Test plan

- [x] All existing tests pass (690+ tests)
- [x] New test for line-range URI format
(`test_multiple_block_uris_use_line_ranges`)
- [x] Content mismatch detection tests in CLI
- [x] Files outside config patterns are ignored by LSP
- [x] TypeScript schema diagnostics appear at correct line positions
(`test_typescript_schema_diagnostics_have_correct_line_offset`)
- [x] TypeScript fix positions verified via
`test_typescript_fix_positions_account_for_block_offset`
- [x] `cargo fmt && cargo clippy && cargo test` all pass


---------

+++ crates/hir/src/lib.rs
@@ -606,7 +606,7 @@ pub struct FragmentNameInfo {
     /// The text range of the fragment name
     pub name_range: TextRange,
     /// For embedded GraphQL: line offset of the block (0-indexed)
-    pub block_line_offset: Option<usize>,
+    pub block_line_offset: Option<u32>,
     /// For embedded GraphQL: byte offset of the block in the original file
     pub block_byte_offset: Option<usize>,
     /// For embedded GraphQL: source text of the block
@@ -649,7 +649,7 @@ pub struct OperationNameInfo {
     /// The text range of the operation name
     pub name_range: Option<TextRange>,
     /// For embedded GraphQL: line offset of the block (0-indexed)
-    pub block_line_offset: Option<usize>,
+    pub block_line_offset: Option<u32>,
     /// For embedded GraphQL: byte offset of the block in the original file
     pub block_byte_offset: Option<usize>,
     /// For embedded GraphQL: source text of the block
